move BitcoinV2::TxOut to modules/output_builder
