+++ rust/tw_bitcoin/src/modules/mod.rs
@@ -2,3 +2,4 @@ mod input_builder;
 mod output_builder;
 
 pub use input_builder::InputBuilder;
+pub use output_builder::OutputBuilder;
